(PC-37106)[PRO] chore: move indiv. offer summary price categories screen + summary entry pages in appropriate folder (new arch)
