+++ pro/src/components/IndividualOffer/IndividualOfferLayout/OfferPublicationEdition/OfferPublicationEditionForm/OfferPublicationEditionForm.tsx
@@ -8,9 +8,9 @@ import { SelectOption } from 'commons/custom_types/form'
 import { FORMAT_HH_mm, formatShortDateForInput } from 'commons/utils/date'
 import { getLocalDepartementDateTimeFromUtc } from 'commons/utils/timezone'
 import { MandatoryInfo } from 'components/FormLayout/FormLayoutMandatoryInfo'
-import { getPublicationHoursOptions } from 'components/IndividualOffer/SummaryScreen/EventPublicationForm/EventPublicationForm'
-import { EventPublicationFormValues } from 'components/IndividualOffer/SummaryScreen/EventPublicationForm/types'
 import { ScrollToFirstHookFormErrorAfterSubmit } from 'components/ScrollToFirstErrorAfterSubmit/ScrollToFirstErrorAfterSubmit'
+import { getPublicationHoursOptions } from 'pages/IndividualOfferSummary/IndividualOfferSummary/components/EventPublicationForm/EventPublicationForm'
+import { EventPublicationFormValues } from 'pages/IndividualOfferSummary/IndividualOfferSummary/components/EventPublicationForm/types'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
 import { DialogBuilder } from 'ui-kit/DialogBuilder/DialogBuilder'
