+++ pro/src/components/IndividualOffer/IndividualOfferLayout/OfferPublicationEdition/OfferPublicationEditionForm/validationSchema.ts
@@ -5,7 +5,7 @@ import {
   bookingAllowedTimeValidationSchema,
   publicationDateValidationSchema,
   publicationTimeValidationSchema,
-} from 'components/IndividualOffer/SummaryScreen/EventPublicationForm/validationSchema'
+} from 'pages/IndividualOfferSummary/IndividualOfferSummary/components/EventPublicationForm/validationSchema'
 
 export const validationSchema = yup.object().shape({
   publicationMode: yup.string<'now' | 'later'>().required().nullable(),
