+++ pro/src/pages/IndividualOfferSummary/IndividualOfferSummary/components/EventPublicationForm/__specs__/validationSchema.spec.ts
