+++ pro/src/pages/IndividualOfferSummary/IndividualOfferSummary/components/OfferSection/OfferSection.module.scss
