+++ pro/src/pages/IndividualOfferSummary/IndividualOfferSummary/components/StockSection/RecurrenceSection/RecurrenceSection.tsx
