+++ pro/src/pages/IndividualOfferSummary/IndividualOfferSummary/components/StockSection/StockSection.module.scss
