+++ pro/src/pages/IndividualOfferSummary/IndividualOfferSummary/components/StockSection/StockSection.tsx
