+++ pro/src/pages/IndividualOfferSummary/IndividualOfferSummary/components/StockSection/__specs__/StockSection.spec.tsx
