+++ pro/src/pages/IndividualOfferSummary/IndividualOfferSummary/components/StockSection/__specs__/getStockWarningText.spec.tsx
