+++ pro/src/pages/IndividualOfferSummary/components/PriceCategoriesSection/PriceCategoriesSection.module.scss
