+++ pro/src/pages/IndividualOfferSummary/components/PriceCategoriesSection/PriceCategoriesSection.tsx
